Refactored the entire program.
Implimented the ChatInterface class, which will allow the robot to be implimented for any chat program (such as MSN) abstractly.
Implimented this change in all of the modules. Modules are now simply hook functions, which take an interface as an argument.
Added the SkypeInterface, so that the program still works with skype.