Many changes:
Added the ComHook class, which replaced separated arrays for the hook name and functions with a dict of hook classes.
AddHook is now redundant. ComHook's constructor does exactly the same thing.
Hooks can now be set to only be called for certain message types, e.g. SENT.
LastMessages() method was added to ChatInterface.