Lots of changes:
Reorganised the module. eval and run now have separate handlers, and separate thread objects.
!eval renamed to !ev
Eval now evaluates inline expressions in messages, designated by {"curley brackets"}.
An example of this would be !ev There are {60*60*24} seconds in a day.
Eval should resort to standard replies if it cannot edit a message. In this case, the function works as before.