Redone a lot of the back end.
Modules are now loaded dynamically in __init__.py. A list of modules in data/modules.txt is loaded on startup.
skype, irc, and msn are now modules on par with any other module.
Modules can be loaded/unloaded on the fly with !load and !unload.
Modules now manage their own setup (I basically moved all the file loading stuff in to the modules). If their startup fails, the error is printed and module loading simply continues.